requirements.txt
contourpy==1.3.1
cycler==0.12.1
et_xmlfile==2.0.0
filterpy==1.4.5
fonttools==4.55.3
joblib==1.5.1
kiwisolver==1.4.8
llvmlite==0.44.0
matplotlib==3.10.0
mplcursors==0.6
numba==0.61.2
numpy==2.2.1
openpyxl==3.1.5
packaging==24.2
pandas==2.3.1
pillow==11.1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.7.1
scipy==1.16.0
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.2

PyYAML~=6.0.2
tqdm~=4.67.1
plotly~=6.2.0